Add workaround for a buffer overflow error in tandem-test
Address sanitizer points to 'bits/string_fortified.h', which leads
to some problem related to the use of _FORTIFY_SOURCE macro.
The source code is not maintained by upstream anymore. Disabled the
macro as a workaround. Note that the macro must be disabled in both
build() and check() for the tandem-test to pass.